feat: 디코딩 관련 버그 수정 및 코드 리팩토링

+++ OpenMarket/OpenMarket/Model/ProductData.swift
@@ -6,7 +6,6 @@
 //
 
 import Foundation
-import UIKit
 
 struct ProductData: Codable {
     let id: Int
@@ -14,8 +13,8 @@ struct ProductData: Codable {
     let name: String
     let thumbnail: String
     let currency: String
-    let price: Int
-    let bargainPrice: Int
+    let price: Double
+    let bargainPrice: Double
     let discountedPrice: Int
     let stock: Int
     let createdAt: String
